Add files via upload
She predicts whether customer in question will subscribe to the product or not based on previous campaign data which has been made available to her. Dataset contains customer characteristics , campaign characteristics, previous campaign information as well as whether customer ended up subscribing to the product or not as a result of that campaign.